using omp in view

+++ View.h
@@ -22,7 +22,7 @@ namespace SBody {
 	  protected:
 		const size_t pixel;
 		std::vector<std::array<double, 9>> initials;
-		std::vector<std::vector<int>> screen;
+		std::vector<std::vector<double>> screen;
 
 	  public:
 		camera(size_t pixel, double viewAngle, double r, double theta, double tFinal, size_t duration = 300, size_t frame = 30);
